New functionality: cmdline interface script
You can now start the server and pass operations to it through the remapp.exs script. This
required modifying the server to send the script's process a :ready message if or when it
is ready. The script itself will register as a local node so communication is made
possible between the server and the script process. When the script receives the message,
it will pass the server options specified by the user, like the csv file path or to delete
the stored data.